requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
waitress==3.0.0
pytest==8.0.0
playwright==1.42.0
